+++ views/mypage/basic/component/LineChart.vue
@@ -56,7 +56,3 @@
         }
     }
 </script>
-
-<style scoped>
-
-</style>
